+++ packages/core/lib/combo/settings/data.ts
@@ -1348,6 +1348,14 @@ export const SETTINGS = [{
   description: 'Enables CAMC for shuffled Cows',
   cond: (x: any) => x.csmc !== 'never',
 }, {
+  key: 'csmcGrassPot',
+  name: 'Grass/Pot CAMC',
+  category: 'main.misc',
+  type: 'boolean',
+  default: false,
+  description: 'Enables CAMC for shuffled Grass and Pots. This is still affected by the Stone of Agony setting.',
+  cond: (x: any) => x.shuffleGrassOot || x.shuffleGrassMm || x.shufflePotsOot || x.shufflePotsMm,
+}, {
   key: 'blastMaskCooldown',
   name: 'Blast Mask Cooldown',
   category: 'main.misc',